python/src/main.py

# well the idea is that for their random config we just steal the returned values and return accordingly
# print(dir())

# can you check if we can steal their random config from memory since it's the same ruuntime??


def actual_initial_config():
    return dict({
      'white': [(3,4), (4,3)],
      'black': [(3,3), (4,4)]
    })

class Game:
    def __init__(self, config: dict = None):
        if config is None:
            self.config: dict = actual_initial_config()
        else:
            self.config: dict = config

    def get_board(self):
        return Board(self.config)


class Board(Game):

    def __init__(self, config: dict):
        super().__init__(config)

    def get_position(self, *args):
        position: tuple = args
        if position in self.config['white']:
            return 'white'
        elif position in self.config['black']:
            return 'black'
        else:
            return "blank"
